[BEWARE] v0.1.2: iOS implementation PoC
iOS implementation of the library. Beware: (1) it slightly alters the
native<>js bridging, thus breaking the Java version, which should be
upgraded to work right again; (2) it also not 100% finished, as it
misses a correct implementation of native<>js message routing at JS side
(currently it works right only with a single stream operating at a
time).