Hotfix: attempt to fix memory leaks & corruption in case of casual use
The previous version was wide open to memory corruption and leaks in
case the user did not wait for async start() operation and did not take
care to track the audio stream state before calling its function. This
is a rapid attempt to fix these problems by making all methods wait for
intialization promise.